separate test cases out, add better comments
