+++ account-compression/sdk/package.json
@@ -45,11 +45,13 @@
     "run-tests": "jest tests --detectOpenHandles",
     "run-tests:events": "jest tests/events --detectOpenHandles",
     "run-tests:accounts": "jest tests/accounts --detectOpenHandles",
+    "run-tests:instructions": "jest tests/instructions --detectOpenHandles",
     "run-tests:e2e": "jest accountCompression.test.ts --detectOpenHandles",
     "test:events": "start-server-and-test start-validator http://127.0.0.1:8899/health run-tests:events",
     "test:accounts": "start-server-and-test start-validator http://127.0.0.1:8899/health run-tests:accounts",
     "test:e2e": "start-server-and-test start-validator http://127.0.0.1:8899/health run-tests:e2e",
     "test:merkle-tree": "jest tests/merkleTree.test.ts --detectOpenHandles",
+    "test:instructions": "start-server-and-test start-validator http://127.0.0.1:8899/health run-tests:instructions",
     "test": "start-server-and-test start-validator http://127.0.0.1:8899/health run-tests"
   },
   "dependencies": {
