+++ account-compression/sdk/tests/merkleTree.test.ts
@@ -1,34 +1,32 @@
-import { assert } from "chai";
-import * as crypto from "crypto";
+import { assert } from 'chai';
+import * as crypto from 'crypto';
 
-import { emptyNode, MerkleTree } from "../src";
+import { emptyNode, MerkleTree } from '../src';
 
-describe("MerkleTree tests", () => {
-  it("Check constructor equivalence for depth 2 tree", () => {
-    const leaves = [
-      crypto.randomBytes(32),
-      crypto.randomBytes(32),
-      crypto.randomBytes(32),
-    ];
-    const rawLeaves = leaves.concat(emptyNode(0));
-    const merkleTreeRaw = new MerkleTree(rawLeaves);
-    const merkleTreeSparse = MerkleTree.sparseMerkleTreeFromLeaves(leaves, 2);
+// This is used to test the MerkleTree helper class that we
+// export to help with the creation and debugging of proofs
+describe('MerkleTree tests', () => {
+    it('Check constructor equivalence for depth 2 tree', () => {
+        const leaves = [crypto.randomBytes(32), crypto.randomBytes(32), crypto.randomBytes(32)];
+        const rawLeaves = leaves.concat(emptyNode(0));
+        const merkleTreeRaw = new MerkleTree(rawLeaves);
+        const merkleTreeSparse = MerkleTree.sparseMerkleTreeFromLeaves(leaves, 2);
 
-    assert(merkleTreeRaw.root.equals(merkleTreeSparse.root));
-  });
+        assert(merkleTreeRaw.root.equals(merkleTreeSparse.root));
+    });
 
-  const TEST_DEPTH = 14;
-  it(`Check proofs for 2^${TEST_DEPTH} tree`, () => {
-    const leaves: Buffer[] = [];
-    for (let i = 0; i < 2 ** TEST_DEPTH; i++) {
-      leaves.push(crypto.randomBytes(32));
-    }
-    const merkleTree = new MerkleTree(leaves);
+    const TEST_DEPTH = 14;
+    it(`Check proofs for 2^${TEST_DEPTH} tree`, () => {
+        const leaves: Buffer[] = [];
+        for (let i = 0; i < 2 ** TEST_DEPTH; i++) {
+            leaves.push(crypto.randomBytes(32));
+        }
+        const merkleTree = new MerkleTree(leaves);
 
-    // Check proofs
-    for (let i = 0; i < leaves.length; i++) {
-      const proof = merkleTree.getProof(i);
-      assert(MerkleTree.verify(merkleTree.getRoot(), proof));
-    }
-  });
+        // Check proofs
+        for (let i = 0; i < leaves.length; i++) {
+            const proof = merkleTree.getProof(i);
+            assert(MerkleTree.verify(merkleTree.getRoot(), proof));
+        }
+    });
 });
